EXPERIMENTAL: use TunnelServer=yes
Use TunnelServer=yes tinc.conf option. With this option set the local
tinc daemon will not accept nodes and subnets from the other tinc
daemons and will only use the settings in the configfiles created
by the chaosvpn program.
With this setting it should be much more difficult for a rogue user to
inject unwanted (aka not in central configfile) routes or to hijack
subnets from other users.