Testing: tinc 1.0.12+git connect only to primary nodes
With the StrictSubnets=yes from tinc 1.0.12+git it works that we only connect
to a selected subpart of the available nodes, and let the tinc peer2peer
feature provide use connectivity to the rest.

This saves all but the primary nodes from a lot of tcp sessions and
reconnect-churn.

For redundancy there should be at least 4-6 primary nodes in the central
config file, one dead primary must not take the network down.